added annotations and various changes to make PUT command actually work with deposits
a little memo: having entity in another entity (for one-to-many &
many-to-one relation with set of said entities, for example) may end up
in infinite json recursion problem. One should use @JsonManagedReference
and @JsonBackReference or @JsonIgnore annotations to solve this. Another
way is to use DTO classes to explicitly define fields you want to get
serialized for the response.